requirements.txt
pybit
python-dotenv
cryptography~=42.0.5